Take token as parameter to pop_used rather than returning it.
This makes more sense; if the device uses a different set of buffers
than the one we were expecting then we shouldn't just go ahead and pop
it, so we return an error instead.

The new peek_used method can be used to check which token is next to pop
and decide whether to handle it.