Use more flexible virtqueue layout when possible.
Only the legacy MMIO transport requires the three parts of the
virtqueue to be contiguous. In other cases we can allocate the regions
written by the driver separately from the region written by the device.
This will reduce the size of the contiguous region allocated, which may
reduce memory fragmentation.