Pass closure to handle packet.
This avoids the need to know which buffer to copy the body into before
seeing the packet header, which will be necessary for handling multiple
connections.

It changes the behaviour of poll_rx_queue (and hence poll_recv) to
immediately return Ok(None) if the first packet in the RX queue is not
relevant, rather than continuing to look for others. This may mean that
poll_recv needs to be called more times than before.