[vsock] Fix credit update in recv() to ensure message transimission
This PR addresses a bug in the vsock connection manager. Prior to
the PR, the guest doesn't update the host after it processes the
local buffer. As a result, when the receiving message length
reaches its buffer capacity, the host could mistakenly stop
sending the message, unaware that the guest still had available
buffer.
This fix ensures that the credit is correctly updated processing
the local buffer, allowing for continuous message transmission
between the host and the guest.