[vsock] Add method to send credit update manually (#110)
* [vsock] Fix credit update in recv() to ensure message transimission

This PR addresses a bug in the vsock connection manager. Prior to
the PR, the guest doesn't update the host after it processes the
local buffer. As a result, when the receiving message length
reaches its buffer capacity, the host could mistakenly stop
sending the message, unaware that the guest still had available
buffer.
This fix ensures that the credit is correctly updated processing
the local buffer, allowing for continuous message transmission
between the host and the guest.

* Update test

* Minor adjustment to retrigger CI

* Minor change to retrigger CI

* Adjust format in examples/aarch64

* Add API to update credit

* Add API is_recv_buffer_empty() in ConnectionManager

* Change is_empty to available

---------

Co-authored-by: Alice Wang <aliceywang@google.com>