queue: check queue size at compile time
Emit a compile-time error if a queue size is invalid, instead of a
runtime error. This allows catching mistakes earlier, since the queue
size is always defined at compile time.

Unfortunately this means we must remove the corresponding test, as the
only way to test a compile-time failure (without using external
crates) is via doctests, which do not have access to private items
(which is the case for VirtQueue).

Signed-off-by: Carlos López <carlos.lopez@suse.com>