Add DeviceDma type and DmaMemory trait
This commit adds a DeviceDma which maps and unmaps memory analogous to the
existing driver-side Dma type. It also adds a DmaMemory trait and implements it
for Dma and DeviceDma to allow using either in VirtQueueLayout. Finally it
imports DmaMemory where its methods are used. This should not be a breaking
change since Dma is not exported by this crate.