PCI: disable I/O and memory decode during BAR sizing
Reprogramming the BAR address to an all-1s value when attempting to
decipher which bits are writable can result in the PCI device actually
intercepting memory or I/O accesses to the resulting address. To avoid
this, disable the I/O and memory bits in the control register during the
BAR sizing process, as recommended by the PCI specification.

Fixes GitHub issue #176.